models/requirements.txt
tensorflow-cpu==2.9.0
numpy==1.23.0
pandas==1.4.3
opencv-python==4.8.1.78
matplotlib==3.5.2
scikit-learn==1.1.1
tensorflowjs==3.18.0
seaborn==0.12.0
pillow==9.3.0 